Scope ReelProtein test dependency stubs
